v0.9.17
Code formatting, menus finishing, card upgrades & traits fixes.

+++ Game/Cards/Internal/Upgrades/FieldCardStatUpgradeException.cs
@@ -7,7 +7,7 @@ namespace Game.Cards
 	/// </summary>
     public class FieldCardStatUpgradeException : Exception
 	{
-		public FieldCardStatUpgradeException(string statName, int upgradePoints, float statPointsShare, float traitPointsShare)
+		public FieldCardStatUpgradeException(string statName, float upgradePoints, float statPointsShare, float traitPointsShare)
 			: base($"Stat \'{statName}\' caused thread block during card upgrade.\n" +
 				   $"upgradePoints: {upgradePoints}, statPointsShare: {statPointsShare}, traitPointsShare: {traitPointsShare}") { }
 	}
