+++ Game/Cards/Internal/Upgrades/FieldCardTraitUpgradeException.cs
@@ -7,7 +7,7 @@ namespace Game.Cards
 	/// </summary>
     public class FieldCardTraitUpgradeException : Exception
 	{
-		public FieldCardTraitUpgradeException(string traitId, int upgradePoints, float statPointsShare, float traitPointsShare)
+		public FieldCardTraitUpgradeException(string traitId, float upgradePoints, float statPointsShare, float traitPointsShare)
 			: base($"Trait \'{traitId}\' caused thread block during card upgrade.\n" +
 				   $"upgradePoints: {upgradePoints}, statPointsShare: {statPointsShare}, traitPointsShare: {traitPointsShare}") { }
 	}
